add threeJS libs and update data
